kageco-273 freee取引データ送信処理のリファクタリング
freeeへの取引データ送信処理をリファクタリングし、簡素化とコードの分離を実施。主な対応内容は以下のとおり:
- 不要な複数レコード対応ロジックを削除し、単一レコードに対応する処理に変更
- `submitDeals`を`submitDealActions`へ変更
- データ変換ロジック`convertToDealData`を専用ファイルに分離
- `findFreeeAuth`を活用し、認証情報の適切な取得を実装
- 冗長なエラーハンドリングコードを削除し簡略化
